Rename `bayou-mocha` -> `testing-server`.
This admits the imminent possibility of a corresponding `testing-client`.
