+++ local-modules/app-setup/package.json
@@ -12,12 +12,12 @@
 
     "api-common": "local",
     "api-server": "local",
-    "bayou-mocha": "local",
     "client-bundle": "local",
     "doc-common": "local",
     "doc-server": "local",
     "see-all": "local",
     "see-all-server": "local",
-    "server-env": "local"
+    "server-env": "local",
+    "testing-server": "local"
   }
 }
