+++ local-modules/testing-server/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "bayou-mocha",
+  "name": "testing-server",
   "version": "1.0.0",
   "main": "main.js",
