+++ server/package.json
@@ -22,12 +22,12 @@
     "source-map-support": "^0.4.2",
 
     "app-setup": "local",
-    "bayou-mocha": "local",
     "client-bundle": "local",
     "dev-mode": "local",
     "hooks-server": "local",
     "see-all": "local",
     "see-all-server": "local",
-    "server-env": "local"
+    "server-env": "local",
+    "testing-server": "local"
   }
 }
